explain issue process for ArchUnit-Examples
Since ArchUnit-Examples is an autogenerated version from the `archunit-examples` submodule of the main project, there is little use to direct issues here. Examples have to be adjusted upstream and questions regarding ArchUnit should also be directed there. I have also as of now deactivated the possibility to open issues in this repo.